requirements.txt
fastapi
uvicorn
pydantic
httpx
pymongo
python-dotenv
dnspython